requirements.txt
absl-py==2.1.0
accelerate==0.26.0
aimet-torch==1.33.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1667325728359/work
async-timeout==5.0.1
attrs==24.3.0
autopep8==2.3.1
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backoff==2.2.1
backports-datetime-fromisoformat==2.0.3
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beautifulsoup4==4.13.0b2
bleach==6.2.0
bokeh==3.2.2
botocore==1.35.90
ccimport==0.4.4
certifi==2024.12.14
charset-normalizer==3.4.1
clarabel==0.9.0
click==8.1.8
cloudpickle==3.1.1
cmake==3.26.0
colorcet==3.1.0
coloredlogs==15.0.1
contextlib2==21.6.0
contourpy==1.3.1
cumm-cu117==0.4.11
cumm-cu120==0.4.11
cvxpy==1.6.0
cycler==0.12.1
cylp==0.92.3
dataclasses==0.6
datasets==3.2.0
debugpy @ file:///home/builder/ci_310/debugpy_1640789504635/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
deprecation==2.1.0
diffusers==0.32.2
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
einops==0.8.0
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
exceptiongroup==1.2.2
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
fastapi==0.115.7
filelock==3.16.1
fire==0.7.0
flatbuffers==24.12.23
fonttools==4.55.4
frozenlist==1.5.0
fsspec==2024.9.0
gdown==4.7.1
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.42
grpcio==1.69.0
h11==0.14.0
h5py==3.12.1
holoviews==1.18.3
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
hvplot==0.9.2
hypothesis==6.124.2
idna==3.10
importlib_metadata==8.5.0
interegular==0.3.3
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1655241617783/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1653754916420/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1659959867326/work
Jinja2==3.0.3
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1652061014773/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1668030830865/work
kiwisolver==1.4.8
lark==1.2.2
linkify-it-py==2.0.3
llvmlite==0.44.0
lm-format-enforcer==0.10.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistral_common==1.5.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==3.4.2
ninja==1.11.1.3
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
onnx==1.14.1
onnxruntime==1.15.1
onnxruntime_extensions==0.13.0
onnxscript==0.1.0.dev20240402
onnxsim==0.4.36
openai==1.60.0
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
osqp==0.6.7.post3
outlines==0.0.46
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.5.3
panel==1.3.8
param==2.2.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
partial-json-parser==0.2.1.1.post5
pccm==0.4.16
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1657729053205/work
portalocker==3.1.1
prettytable==3.11.0
progressbar2==4.5.0
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1667565496306/work
propcache==0.2.1
protobuf==3.20.2
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1667885877572/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==18.1.0
pybind11==2.13.6
pycodestyle==2.12.1
pycountry==24.6.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1660666458521/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv==1.0.1
python-utils==3.9.1
pytz==2024.2
pyviz_comms==3.0.4
PyYAML==6.0.2
pyzmq @ file:///home/builder/ci_310/pyzmq_1640794096902/work
qai-hub==0.20.1
-e git+https://github.com/quic/ai-hub-models.git@9857d3ffe0df7946040aec35628e1f6582b03095#egg=qai_hub_models
qdldl==0.1.7.post5
ray==2.41.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.10.4
safetensors==0.4.5
schema==0.7.5
scikit-learn==1.1.3
scipy==1.8.1
scs==3.2.7.post2
semver==3.0.2
sentencepiece==0.2.0
setuptools-scm==8.1.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
spconv-cu117==2.3.6
spconv-cu120==2.3.6
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1668260892666/work
starlette==0.45.2
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.4
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.2.1
torch==2.2.2
torchvision==0.17.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1666788589303/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1666115969632/work
transformers==4.40.0
triton==2.2.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
websockets==14.2
Werkzeug==3.1.3
xformers==0.0.27.post2
xxhash==3.5.0
xyzservices==2025.1.0
yarl==1.18.3
zipp==3.21.0
